+++ cmd/local/jvmcollect/jfr.go
@@ -18,6 +18,7 @@ package jvmcollect
 import (
 	"bytes"
 	"fmt"
+	"strings"
 	"time"
 
 	"github.com/dremio/dremio-diagnostic-collector/cmd/local/conf"
@@ -31,13 +32,16 @@ func RunCollectJFR(c *conf.CollectConf) error {
 	if err := ddcio.Shell(&w, fmt.Sprintf("jcmd %v VM.unlock_commercial_features", c.DremioPID())); err != nil {
 		simplelog.Warningf("Error trying to unlock commercial features %v. Note: newer versions of OpenJDK do not support the call VM.unlock_commercial_features. This is usually safe to ignore", err)
 	}
+
 	simplelog.Debugf("node: %v - jfr unlock commercial output - %v", c.NodeName(), w.String())
 
 	w = bytes.Buffer{}
+	// this is effectively a no op unless there is an existing recording running
 	if err := ddcio.Shell(&w, fmt.Sprintf("jcmd %v JFR.stop name=\"DREMIO_JFR\"", c.DremioPID())); err != nil {
 		simplelog.Debugf("attempting to stop existing JFR failed, but this is usually expected: '%v' -- output: '%v'", err, w.String())
-	} else {
-		simplelog.Warningf("JFR named DREMIO_JFR was running on PID %v and has been stopped so we could get a fresh collection: '%v'", c.DremioPID(), w.String())
+	}
+	if strings.Contains(w.String(), "Stopped recording \"DREMIO_JFR\"") {
+		simplelog.Warningf("stopped a JFR recording named \"DREMIO_JFR\"")
 	}
 
 	w = bytes.Buffer{}